+++ lib/simple_form/form_builder.rb
@@ -11,6 +11,8 @@ module SimpleForm
       :update => :edit
     }
 
+    ATTRIBUTE_COMPONENTS = [:html5, :min_max, :maxlength, :placeholder, :pattern, :readonly]
+
     extend MapType
     include SimpleForm::Inputs
 
@@ -134,10 +136,10 @@ module SimpleForm
     #
     def input_field(attribute_name, options={})
       options = options.dup
-      options[:input_html] = options.except(:as, :collection, :label_method, :value_method)
+      options[:input_html] = options.except(:as, :collection, :label_method, :value_method, *ATTRIBUTE_COMPONENTS)
       options = @defaults.deep_dup.deep_merge(options) if @defaults
 
-      SimpleForm::Wrappers::Root.new([:min_max, :maxlength, :placeholder, :pattern, :readonly, :input], :wrapper => false).render find_input(attribute_name, options)
+      SimpleForm::Wrappers::Root.new(ATTRIBUTE_COMPONENTS + [:input], :wrapper => false).render find_input(attribute_name, options)
     end
 
     # Helper for dealing with association selects/radios, generating the